+++ R/simsum.R
@@ -21,12 +21,14 @@
 #' Useful for non-Wald type estimators (e.g. bootstrap).
 #' Defaults to `NULL`, where Wald-type confidence intervals based on the provided SEs are calculated for coverage; otherwise, it can be a numeric vector (for fixed confidence intervals) or a vector of strings that identify columns in `data` with replication-specific lower and upper limits.
 #' See `vignette("E-custom-inputs", package = "rsimsum")` for more details.
+#' @param df Can be used to specify that a column containing the replication-specific number of degrees of freedom that will be used to calculate confidence intervals for coverage (and bias-eliminated coverage) assuming t-distributed critical values (rather than normal theory intervals).
+#' See `vignette("E-custom-inputs", package = "rsimsum")` for more details.
 #' @param dropbig Specifies that point estimates or standard errors beyond the maximum acceptable values should be dropped. Defaults to `FALSE`.
 #' @param x Set to `TRUE` to include the `data` argument used to calculate summary statistics (i.e. after pre-processing the input dataset e.g. removing values deemed too large via the `dropbig` argument) as a slot. Calling `simsum` with `x = TRUE` is required to produce zipper plots. The downside is that the size of the returned object increases considerably, therefore it is set to `FALSE` by default.
 #' @param control A list of parameters that control the behaviour of `simsum`. Possible values are:
 #' * `mcse`, whether to calculate Monte Carlo standard errors. Defaults to `TRUE`;
 #' * `level`, the significance level used for coverage, bias-eliminated coverage, and power. Defaults to 0.95;
-#' * `df`, whether to use robust critical values from a t distribution with `df` degrees of freedom when calculating coverage, bias-eliminated coverage, and power. Defaults to `NULL`, in which case a Gaussian distribution is used;
+#' * `power_df`, whether to use robust critical values from a t distribution with `power_df` degrees of freedom when calculating power. Defaults to `NULL`, in which case a Gaussian distribution is used;
 #' * `na.rm`, whether to remove point estimates or standard errors where either (or both) is missing. Defaults to `TRUE`;
 #' * `char.sep`, a character utilised when splitting the input dataset `data`. Generally, this should not be changed;
 #' * `dropbig.max`, specifies the maximum acceptable absolute value of the point estimates, after standardisation. Defaults to 10;
@@ -53,6 +55,7 @@ simsum <- function(data,
                    ref = NULL,
                    by = NULL,
                    ci.limits = NULL,
+                   df = NULL,
                    dropbig = FALSE,
                    x = FALSE,
                    control = list()) {
@@ -65,6 +68,7 @@ simsum <- function(data,
   checkmate::assert_string(x = se, null.ok = TRUE, add = arg_checks)
   checkmate::assert_string(x = methodvar, null.ok = TRUE, add = arg_checks)
   checkmate::assert_string(x = ref, null.ok = TRUE, add = arg_checks)
+  checkmate::assert_string(x = df, null.ok = TRUE, add = arg_checks)
   # 'true' must be a single numeric value, or a string that identifies a column in 'data'
   if (!is.null(true)) {
     checkmate::assert_true(x = inherits(x = true, what = c("character", "numeric", "integer")), add = arg_checks)
@@ -80,20 +84,22 @@ simsum <- function(data,
   checkmate::assert_logical(x = x, len = 1, add = arg_checks)
   # 'by' must be a vector of strings; can be NULL
   checkmate::assert_character(x = by, null.ok = TRUE, add = arg_checks)
-  # 'estvarname', 'se' must be in 'data'; all elements of 'by' must be in 'data'; 'methodvar' must be in 'data'
+  # 'estvarname', 'se' must be in 'data'; all elements of 'by' must be in 'data'; 'methodvar' must be in 'data'; 'df' must be in 'data'
   checkmate::assert_subset(x = estvarname, choices = names(data), add = arg_checks)
   checkmate::assert_subset(x = se, choices = names(data), add = arg_checks)
   checkmate::assert_subset(x = by, choices = names(data), add = arg_checks)
   checkmate::assert_subset(x = methodvar, choices = names(data), add = arg_checks)
+  checkmate::assert_subset(x = df, choices = names(data), add = arg_checks)
   # 'ref' must be one of the options in 'methodvar'
   if (!is.null(methodvar)) {
     checkmate::assert_subset(x = ref, choices = as.character(unique(data[[methodvar]])), add = arg_checks)
   }
-  # 'estvarname', 'se', 'methodvar', 'by' must not be any in ('stat', 'est', 'mcse', 'lower', 'upper')
+  # 'estvarname', 'se', 'methodvar', 'by' , 'df' must not be any in ('stat', 'est', 'mcse', 'lower', 'upper')
   checkmate::assert_false(x = (estvarname %in% c("stat", "est", "mcse", "lower", "upper")), add = arg_checks)
   if (!is.null(se)) checkmate::assert_false(x = (se %in% c("stat", "est", "mcse", "lower", "upper")), add = arg_checks)
   if (!is.null(methodvar)) checkmate::assert_false(x = (methodvar %in% c("stat", "est", "mcse", "lower", "upper")), add = arg_checks)
   if (!is.null(by)) checkmate::assert_false(x = any(by %in% c("stat", "est", "mcse", "lower", "upper")), add = arg_checks)
+  if (!is.null(df)) checkmate::assert_false(x = any(df %in% c("stat", "est", "mcse", "lower", "upper")), add = arg_checks)
   # 'ci.limits' must be either a numeric vector of length 2 or a string vector with column names in 'data'
   if (!is.null(ci.limits)) {
     checkmate::assert_true(x = inherits(x = ci.limits, what = c("character", "numeric", "integer")), add = arg_checks)
@@ -107,7 +113,7 @@ simsum <- function(data,
   }
   # 'control' must be a list, with well defined components
   checkmate::assert_list(x = control, add = arg_checks)
-  checkmate::assert_subset(x = names(control), choices = c("mcse", "level", "df", "na.rm", "char.sep", "dropbig.max", "dropbig.semax", "dropbig.robust"), empty.ok = TRUE, add = arg_checks)
+  checkmate::assert_subset(x = names(control), choices = c("mcse", "level", "power_df", "na.rm", "char.sep", "dropbig.max", "dropbig.semax", "dropbig.robust"), empty.ok = TRUE, add = arg_checks)
   checkmate::assert_logical(x = control$mcse, len = 1, null.ok = TRUE, add = arg_checks)
   checkmate::assert_number(x = control$level, lower = 0, upper = 1, null.ok = TRUE, add = arg_checks)
   checkmate::assert_number(x = control$df, null.ok = TRUE, add = arg_checks)
@@ -118,6 +124,9 @@ simsum <- function(data,
   checkmate::assert_logical(x = control$dropbig.robust, len = 1, null.ok = TRUE, add = arg_checks)
   # Report
   if (!arg_checks$isEmpty()) checkmate::reportAssertions(arg_checks)
+
+  ### Only one of 'ci.limits' and 'df' can be specified
+  if (!is.null(ci.limits) & !is.null(df)) stop("Only one of 'ci.limits' and 'df' can be specified.", call. = FALSE)
 
   ### Set control parameters
   control.default <- list(mcse = TRUE, level = 0.95, df = NULL, na.rm = TRUE, char.sep = "~", dropbig.max = 10, dropbig.semax = 100, dropbig.robust = TRUE)
@@ -177,7 +186,7 @@ simsum <- function(data,
       } else {
         empse_ref <- NULL
       }
-      out.in <- .performance(data = data[[i]][[j]], estvarname = estvarname, se = se, true = true, rho = rho[names(data[[i]][j])], empse_ref = empse_ref, ci.limits = ci.limits, control = control)
+      out.in <- .performance(data = data[[i]][[j]], estvarname = estvarname, se = se, true = true, rho = rho[names(data[[i]][j])], empse_ref = empse_ref, ci.limits = ci.limits, df = df, control = control)
       if (!is.null(methodvar)) {
         out.in[[methodvar]] <- unique(data[[i]][[j]][[methodvar]])
       }
@@ -200,9 +209,8 @@ simsum <- function(data,
   obj$methodvar <- methodvar
   obj$ref <- ref
   obj$dropbig <- dropbig
-  if (!is.null(ci.limits)) {
-    obj$ci.limits <- ci.limits
-  }
+  obj$ci.limits <- ci.limits
+  obj$df <- df
   obj$by <- by
   obj$control <- control
   if (x) {