+++ tests/testthat/test-autoplot.R
@@ -328,20 +328,23 @@ testthat::test_that("inferring target", {
 })
 
 testthat::test_that("zip with t critical values", {
-  data("MIsim", package = "rsimsum")
-  data("relhaz", package = "rsimsum")
-  single <- rsimsum::simsum(data = MIsim, estvarname = "b", true = 0.5, se = "se", methodvar = "method", ref = "CC", x = TRUE, control = list(df = 10))
-  multi <- rsimsum::simsum(data = relhaz, estvarname = "theta", true = -0.5, se = "se", methodvar = "model", by = c("n", "baseline"), x = TRUE, control = list(df = 10))
+  data("tt", package = "rsimsum")
+  tt1 <- tt2 <- tt
+  tt1$yo <- "One"
+  tt2$yo <- "Two"
+  doublett <- do.call(rbind.data.frame, list(tt1, tt2))
+
+  single <- rsimsum::simsum(data = tt, estvarname = "diff", true = -1, se = "se", methodvar = "method", x = TRUE, df = "df")
+  multi <- rsimsum::multisimsum(data = doublett, par = "yo", estvarname = "diff", true = c(One = -1, Two = -1), se = "se", methodvar = "method", x = TRUE, df = "df")
   singlesum <- summary(single)
   multisum <- summary(multi)
-  # simsum object, no 'by'
+
   testthat::expect_s3_class(object = autoplot(single, type = "zip"), class = c("gg", "ggplot"))
-  # simsum object, with 'by'
-  testthat::expect_s3_class(object = autoplot(multi, type = "zip"), class = c("gg", "ggplot"))
-  # summary.simsum object, no 'by'
+  testthat::expect_s3_class(object = autoplot(multi, par = "One", type = "zip"), class = c("gg", "ggplot"))
+  testthat::expect_s3_class(object = autoplot(multi, par = "Two", type = "zip"), class = c("gg", "ggplot"))
   testthat::expect_s3_class(object = autoplot(singlesum, type = "zip"), class = c("gg", "ggplot"))
-  # summary.simsum object, with 'by'
-  testthat::expect_s3_class(object = autoplot(multisum, type = "zip"), class = c("gg", "ggplot"))
+  testthat::expect_s3_class(object = autoplot(multisum, par = "One", type = "zip"), class = c("gg", "ggplot"))
+  testthat::expect_s3_class(object = autoplot(multisum, par = "Two", type = "zip"), class = c("gg", "ggplot"))
 })
 
 data("frailty", package = "rsimsum")
@@ -485,4 +488,47 @@ testthat::test_that("Edge cases for rsimsum >= 0.8.0", {
   testthat::expect_error(object = autoplot(s1, type = "lolly", stats = "thetamedian"))
   testthat::expect_error(object = autoplot(s4, type = "lolly", stats = "thetamean"))
   testthat::expect_error(object = autoplot(s4, type = "lolly", stats = "thetamedian"))
+})
+
+### More tests for zip plots...
+testthat::test_that("Testing more zip plot cases", {
+  data("tt", package = "rsimsum")
+
+  s <- simsum(data = tt, estvarname = "diff", se = "se", true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", df = "df", true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", ci.limits = c("lower", "upper"), true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+
+  s <- simsum(data = tt, estvarname = "diff", se = "se", methodvar = "method", true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", methodvar = "method", df = "df", true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", methodvar = "method", ci.limits = c("lower", "upper"), true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+
+  s <- simsum(data = tt, estvarname = "diff", se = "se", by = "dgm", true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", by = "dgm", df = "df", true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", by = "dgm", ci.limits = c("lower", "upper"), true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+
+  s <- simsum(data = tt, estvarname = "diff", se = "se", methodvar = "method", by = "dgm", true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", methodvar = "method", by = "dgm", df = "df", true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", methodvar = "method", by = "dgm", ci.limits = c("lower", "upper"), true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", ci.limits = c(-2, 0), true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", methodvar = "method", ci.limits = c(-2, 0), true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", by = "dgm", ci.limits = c(-2, 0), true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
+  s <- simsum(data = tt, estvarname = "diff", se = "se", methodvar = "method", by = "dgm", ci.limits = c(-2, 0), true = -1, x = TRUE)
+  testthat::expect_s3_class(object = autoplot(s, type = "zip"), class = c("gg", "ggplot"))
 })